Updated libopenmpt to version 0.5.12

+++ Frameworks/OpenMPT/OpenMPT/soundlib/plugins/dmo/Distortion.cpp
@@ -27,7 +27,7 @@ namespace DMO
 // Computes (log2(x) + 1) * 2 ^ (shiftL - shiftR) (x = -2^31...2^31)
 float logGain(float x, int32 shiftL, int32 shiftR)
 {
-	uint32 intSample = static_cast<uint32>(static_cast<int32>(x));
+	uint32 intSample = static_cast<uint32>(static_cast<int64>(x));
 	const uint32 sign = intSample & 0x80000000;
 	if(sign)
 		intSample = (~intSample) + 1;
